// app/src/generated/api/models/UsersPermissionsUser.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: mateja176@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UsersPermissionsUser
 */
export interface UsersPermissionsUser {
    /**
     * 
     * @type {string}
     * @memberof UsersPermissionsUser
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof UsersPermissionsUser
     */
    username: string;
    /**
     * 
     * @type {string}
     * @memberof UsersPermissionsUser
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UsersPermissionsUser
     */
    provider?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UsersPermissionsUser
     */
    confirmed?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UsersPermissionsUser
     */
    blocked?: boolean;
    /**
     * 
     * @type {object}
     * @memberof UsersPermissionsUser
     */
    role?: object;
    /**
     * 
     * @type {Array<object>}
     * @memberof UsersPermissionsUser
     */
    trips?: Array<object>;
}

export function UsersPermissionsUserFromJSON(json: any): UsersPermissionsUser {
    return UsersPermissionsUserFromJSONTyped(json, false);
}

export function UsersPermissionsUserFromJSONTyped(json: any, ignoreDiscriminator: boolean): UsersPermissionsUser {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'username': json['username'],
        'email': json['email'],
        'provider': !exists(json, 'provider') ? undefined : json['provider'],
        'confirmed': !exists(json, 'confirmed') ? undefined : json['confirmed'],
        'blocked': !exists(json, 'blocked') ? undefined : json['blocked'],
        'role': !exists(json, 'role') ? undefined : json['role'],
        'trips': !exists(json, 'trips') ? undefined : json['trips'],
    };
}

export function UsersPermissionsUserToJSON(value?: UsersPermissionsUser | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'username': value.username,
        'email': value.email,
        'provider': value.provider,
        'confirmed': value.confirmed,
        'blocked': value.blocked,
        'role': value.role,
        'trips': value.trips,
    };
}